HIP-62 | Raghavi/Abhijeet | Extract outs organization context building logic
