+++ omod/src/main/java/org/bahmni/module/hip/web/model/OrganizationContext.java
@@ -1,4 +1,4 @@
-package org.bahmni.module.hip.web.service;
+package org.bahmni.module.hip.web.model;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -6,13 +6,16 @@ import org.hl7.fhir.r4.model.Organization;
 import org.openmrs.VisitType;
 
 @Builder
-@Getter
-public class OrgContext {
+public class OrganizationContext {
     private Organization organization;
     private String webUrl;
 
-    public Class getCareContextType() {
+    public Class careContextType() {
         //Hardcoded right now. Should also deal with programType, visit or visitType.
         return VisitType.class;
+    }
+
+    public String webUrl() {
+        return webUrl;
     }
 }
